+++ src/containers/lists/lists.js
@@ -1,7 +1,8 @@
 import { useSelector, useDispatch } from 'react-redux'
 import Layout from 'layouts/with-sidebar'
 import { SideNav } from 'connectors/side-nav/side-nav'
-import { ListsHeader } from 'components/headers/lists-header'
+import { ListsAllHeader } from 'components/headers/lists-header'
+import { EmptyAllLists } from 'components/empty-states/all-lists'
 import { listsItemsSetSortOrder } from './lists.state'
 
 export const Lists = () => {
@@ -23,7 +24,7 @@ export const Lists = () => {
 
       {shouldRender ? (
         <main className="main">
-          <ListsHeader
+          <ListsAllHeader
             sortOrder={sortOrder}
             handleCreateList={handleCreateList}
             handleNewest={handleNewest}
@@ -31,6 +32,8 @@ export const Lists = () => {
           />
 
           {/* List Component */}
+          <EmptyAllLists />
+
         </main>
       ) : null}
     </Layout>